feat: 병렬 라우팅 아닌 예제 렌더링 방식 수정
